Verify JWT token
- Add nonce and verify on callback phase
- Verify `aud` (audience) jwt claim. It should be the same as your
client id.
- Explicitly pass verify options to JWT.decode, including expiration,
not before, iat.
- Not verifying iss because it can be a different value depending on the
type of account that's being logged in. For an Azure AD account, I'm not
sure yet how to verify this value or if it should even *be* verfied. The
docs don't mention anything about verifying this value; they simply
state that you can figure out if the person used a personal account by
matching up the iss value to a constant value that Microsoft uses to
identify that type of account.
- Add rubocop.yml file. I might tweak these options. Mainly didn't want
warnings for line length over 80.

Still to do:

- Add new tests where needed and adapt tests from the other two
  repositores and change as needed.
- Update Docs

Also, this isn't really tested for a client that is not using the
'common' endpoint. I probably won't be spending any time on that part.